refactor: rename cell_types.js to cellTypes.js
