+++ client/src/constants/cellTypes.js
